Shortened Cult Application
So instead of just proposing we replace the application or "fix what's broken", we just have two applications for something Rachel wants to try and might be easier on both cultists to speed up the process and make following up on apps more interesting. The goal here is to shorten the application to essential information and then to have the potential cultist be interviewed, which could also lead into finding their gear. The questions during the interview could either be a separate document or just filling out the non-junk questions on the normal application for the interviewer, which we retain for record-keeping.